+++ CHANGELOG.md
@@ -41,6 +41,7 @@ going to the next item
 - Scrollbars now represent the viewport position instead of the currently selected item position
 - Remote commands now check for `$PID` env variable, meaning `--pid` argument is no longer needed for
 remote commands inside scripts triggered by rmpc
+- `AddToPlaylist` binding handles marked songs rather than only the one under your cursor.
 
 ### Fixed